web_ui/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
pyserial>=3.5
